+++ ext_localconf.php
@@ -102,6 +102,10 @@ $boot = static function () {
         'priority' => 70,
         'class' => \OliverKlee\Seminars\BackEnd\TimeSlotWizard::class,
     ];
+
+    // FAL upgrade wizards
+    $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['ext/install']['update']['seminars_migrateCategoryIconsToFal']
+        = \OliverKlee\Seminars\UpgradeWizards\CategoryIconToFalUpgradeWizard::class;
 };
 
 $boot();